refactor(web): move the stray page modules under features/agents
AGENTS.md places page and domain modules in `src/features`, but three sat at
the root of `src`. Each moves to the directory its consumers already live in:

- `computer-setup` to `features/agents/`, at the top of the domain rather than
  inside `agent-settings/`, because agent settings and agent creation both use it
- `runtime-configuration` to `features/agents/agent-settings/`, beside
  `general-config-form` and its only consumer
- `agent-detail-capabilities` to `features/agents/`

There are no path aliases or barrel files, so every relative import in the moved
files and in their three consumers is rewritten by hand; `tsc` covers the rest.
No behavior changes.
